pyproject.toml
[tool.poetry]
name = "yokoprotocol"
version = "0.1.0"
description = ""
authors = ["Maksim Shushkevich <meshushkevich@edu.hse.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
colorama = "^0.4.6"
termcolor = "^2.4.0"
pystun3 = "^2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
